refactor organization management panels
